[status] Refactor
